Move serde to src
