+++ bindings/go/src/entities/popdata.go
@@ -1,8 +1,15 @@
 package entities
 
-// import (
-// 	"github.com/VeriBlock/alt-integration-cpp/src/base"
-// )
+import (
+	"io"
+
+	veriblock "github.com/VeriBlock/alt-integration-cpp/src"
+)
 
 // PopData ...
 type PopData struct{}
+
+// FromRaw ...
+func (v PopData) FromRaw(r io.Reader) {
+	_, _ = veriblock.ReadArrayOf(r, veriblock.ReadArrayOfFunc)
+}
